MR : Feat: Add UserRepository.java | juyoung => repository

See merge request s09-webmobile3-sub2/S09P12A304!41

+++ BackEnd/ggong/src/main/java/com/a304/ggong/dto/UserDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 public class UserDto {
 
-    private int user_no;
+    private Long user_no;
     private String name;
     private String age_range;
     private String gender;
